Change trigger height to 268mm
Based upon some testing, and adding in something of a "safety margin", I've increased the trigger height for turning off the heater at 268mm.  This is at about the point where the thermal protection sensor temperature starts to slowly climb.

Rather than turning it off too early, or waiting until the absolute last moment to turn the heater off, which may risk triggering a thermal shutdown, we'll turn it off at the latest point before the thermal protection sensor starts to show deviation.

This means that the printer will only have to "coast" for the last 12mm of Z travel.